# front/G-stock/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/stock",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/user/{id}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GrantedAuthority": {
                "type": "object",
                "properties": {
                    "authority": {
                        "type": "string"
                    }
                }
            },
            "Role": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "accountLocked": {
                        "type": "boolean"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "name": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "authorities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GrantedAuthority"
                        }
                    },
                    "accountNonExpired": {
                        "type": "boolean"
                    },
                    "credentialsNonExpired": {
                        "type": "boolean"
                    },
                    "accountNonLocked": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}